requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.35.0
botocore==1.35.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chonkie==1.4.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docling>=2.63.0
docling-core>=2.52.0
docling-ibm-models>=3.10.0
docling-parse>=4.7.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
firecrawl-py==1.2.4
Flask==3.0.3
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_git==0.50.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
langchain==0.3.27
langchain-core==0.3.80
langchain-litellm==0.3.2
langchain-text-splitters==0.3.11
langsmith==0.4.47
lightrag-hku==1.4.9.8
litellm==1.80.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbdime==4.0.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.109.1
openpyxl==3.1.5
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pgvector==0.3.6
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.3.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.1
referencing==0.35.1
regex==2024.9.11
rerankers[flashrank]==0.10.0
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.1.0
setuptools==73.0.1
six==1.16.0
slowapi==0.1.9
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
sympy==1.12
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
#torch==2.9.1+cpu
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.4
yarl==1.11.1
youtube-transcript-api==0.6.3

# Optional: Vision dependencies for advanced document processing
# Uncomment to enable ONNX-based table structure recognition and OCR
# onnxruntime>=1.18.0
# opencv-python-headless>=4.10.0
# paddleocr>=2.7.0
# paddlepaddle>=2.6.0
